+++ tests/test_internal_api_base.py
@@ -41,7 +41,7 @@ def test_no_hardcoded_loopback_left_in_call_sites():
         "src/tools/_common.py",
         "src/cookbook_serve_lifecycle.py",
         "src/builtin_actions.py",
-        "routes/task_routes.py",
+        "routes/task/task_routes.py",
     ):
         text = (root / rel).read_text(encoding="utf-8")
         # Allow it only inside comments; flag any code occurrence.
