+++ tests/test_model_helper_owner_scope.py
@@ -27,7 +27,7 @@ def test_calendar_quick_parse_resolves_with_owner_scope():
 
 
 def test_task_parse_resolves_with_owner_scope():
-    body = _function_source("routes/task_routes.py", "parse_task")
+    body = _function_source("routes/task/task_routes.py", "parse_task")
     assert "user = _owner(request)" in body
     assert 'resolve_endpoint("utility", owner=user or None)' in body
     assert 'resolve_endpoint("default", owner=user or None)' in body
